refactor: 拆分 blocked_state_service 为三个模块
拆分前: blocked_state_service.py 692 行（过大，阻碍维护）

拆分后:
- blocked_state_types.py (54 行): 数据模型 BlockedState, ConsistencyReport
- blocked_state_io.py (227 行): I/O 操作 (read/write body/DB/labels)
- blocked_state_service.py (299 行): 主服务公共 API

拆分保持:
- 所有测试通过 (50 passed)
- 公共 API 不变
- 私有方法委托给 BlockedStateIO

LOC 限制从 700 降到 350
